requirements.txt
blessed==1.19.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
inquirer==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
readchar==2.0.1
requests==2.26.0
six==1.16.0
TogglPy==0.1.2
urllib3==1.26.6
wcwidth==0.2.5
